Move PBEM/PBF panel classes to be co-located with AbstractForumPosterPanel
