+++ game-core/src/main/java/games/strategy/triplea/ui/PoliticsPanel.java
@@ -193,7 +193,7 @@ public class PoliticsPanel extends ActionPanel {
   }
 
   @Override
-  void performDone() {
+  public void performDone() {
     if (!firstRun || youSureDoNothing()) {
       choice = null;
       release();
