+++ game-core/src/main/java/games/strategy/triplea/ui/TechPanel.java
@@ -243,7 +243,7 @@ class TechPanel extends ActionPanel {
   }
 
   @Override
-  void performDone() {
+  public void performDone() {
     techRoll = null;
     release();
   }
